הגדרות צורת הדף לחלונית צפה ובורר מפרשים כתפריט חיפוש נפתח
- הגדרות צורת הדף עברו מדיאלוג מודאלי לחלונית צפה (ContextOverlayPanel) בצד ימין, כמו שאר חלוניות ההגדרות באפליקציה.
- בורר בחירת המפרשים (4 המיקומים) הומר מדיאלוג לתפריט נפתח מעוגן עם חיפוש מובנה (AppDropdownField) וצ'יפים לסינון לפי דורות.
- ניקוי תאימות בפאנל: עקיפות withValues הועברו לטוקנים ב-AppSurfaces (panelSection/Accent) ולצבע onSurfaceVariant, ובחירת תחום שמירת המפרשים עברה מ-RadioGroup ל-AppSegmentedControl.

כולל גם, מסשן עבודה מקביל ובלתי-נפרד בקבצים משותפים: שילוב SplitPaneContentInset במסכי הקריאה (pdf_book, commentators, full_text_search, reading_screen, splited_view).
