+++ lib/search/view/full_text_search_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:otzaria/tabs/models/searching_tab.dart';
 import 'package:otzaria/search/view/tantivy_full_text_search.dart';
+import 'package:otzaria/widgets/layout/split_pane_content_inset.dart';
 import 'package:flutter/material.dart';
 
 class FullTextSearchScreen extends StatelessWidget {
@@ -8,10 +9,13 @@ class FullTextSearchScreen extends StatelessWidget {
   const FullTextSearchScreen({super.key, required this.tab});
   @override
   Widget build(BuildContext context) {
-    return BlocProvider.value(
-      value: tab.searchBloc,
-      child: TantivyFullTextSearch(
-        tab: tab,
+    return Padding(
+      padding: SplitPaneContentInset.of(context),
+      child: BlocProvider.value(
+        value: tab.searchBloc,
+        child: TantivyFullTextSearch(
+          tab: tab,
+        ),
       ),
     );
   }
