+++ test/tabs/reading_screen_regression_test.dart
@@ -1,6 +1,29 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_settings_screens/flutter_settings_screens.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:otzaria/core/focus_repository.dart';
+import 'package:otzaria/models/books.dart';
+import 'package:otzaria/models/pdf_headings.dart';
+import 'package:otzaria/pdf_book/view/pdf_commentators_tab_screen.dart';
+import 'package:otzaria/personal_notes/bloc/personal_notes_bloc.dart';
+import 'package:otzaria/personal_notes/bloc/personal_notes_event.dart';
+import 'package:otzaria/personal_notes/bloc/personal_notes_state.dart';
+import 'package:otzaria/settings/engine/settings_bloc.dart';
+import 'package:otzaria/settings/engine/settings_event.dart';
+import 'package:otzaria/settings/engine/settings_state.dart';
+import 'package:otzaria/tabs/bloc/tabs_bloc.dart';
+import 'package:otzaria/tabs/bloc/tabs_event.dart';
+import 'package:otzaria/tabs/bloc/tabs_state.dart';
+import 'package:otzaria/tabs/models/combined_tab.dart';
+import 'package:otzaria/tabs/models/pdf_commentators_tab.dart';
+import 'package:otzaria/tabs/models/pdf_tab.dart';
+import 'package:otzaria/tabs/reading_screen.dart';
+import 'package:otzaria/widgets/layout/adaptive_side_pane.dart';
+import 'package:provider/provider.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
+
+import '../helpers/memory_settings_cache.dart';
 
 /// טסטי רגרסיה לתיקונים בקובץ `lib/tabs/reading_screen.dart`:
 ///
@@ -247,6 +270,99 @@ void main() {
     );
   });
 
+  group('reading_screen — שוליים חיצוניים בתצוגה זה-לצד-זה', () {
+    // רגרסיה: הטקסט בתצוגה המפוצלת נצמד לדופן החלון. התיקון מזריק שוליים של 12
+    // (רוחב המפריד) לתוך *תוכן* כל חלונית בלבד, דרך SplitPaneContentInset,
+    // כך שהחלונית עצמה נשארת צמודה לדופן — ידית פתיחת החלונית
+    // (Positioned(left:0)) יושבת בדיוק על הדופן, בעוד הטקסט שומר 12px.
+    const dividerWidth = 12.0;
+
+    setUpAll(() async {
+      await Settings.init(cacheProvider: MemorySettingsCache());
+    });
+
+    testWidgets(
+      'ReadingScreen עם CombinedTab: החלוניות צמודות לדופן והתוכן נשמר 12px פנימה',
+      (tester) async {
+        tester.view.physicalSize = const Size(1600, 900);
+        tester.view.devicePixelRatio = 1.0;
+        addTearDown(tester.view.resetPhysicalSize);
+        addTearDown(tester.view.resetDevicePixelRatio);
+
+        final combined = CombinedTab(
+          rightTab: _commentaryTab('ימין'),
+          leftTab: _commentaryTab('שמאל'),
+        );
+        addTearDown(combined.dispose);
+
+        final tabsBloc = _FakeTabsBloc(
+          TabsState(tabs: [combined], currentTabIndex: 0),
+        );
+        addTearDown(tabsBloc.close);
+
+        await tester.pumpWidget(
+          MaterialApp(
+            home: Directionality(
+              // כמו באפליקציה האמיתית (locale he) — כדי שהחלפת start/end
+              // בשוליי SplitPaneContentInset תתגלה בכיוון הייצור.
+              textDirection: TextDirection.rtl,
+              child: MultiBlocProvider(
+                providers: [
+                  BlocProvider<SettingsBloc>.value(value: _FakeSettingsBloc()),
+                  BlocProvider<PersonalNotesBloc>.value(
+                    value: _FakePersonalNotesBloc(),
+                  ),
+                  BlocProvider<TabsBloc>.value(value: tabsBloc),
+                  Provider<FocusRepository>.value(value: FocusRepository()),
+                ],
+                child: const ReadingScreen(),
+              ),
+            ),
+          ),
+        );
+        await tester.pump();
+
+        final panes = find.byType(PdfCommentatorsTabScreen);
+        expect(panes, findsNWidgets(2));
+
+        // כל חלונית נצמדת לדופן החלון — הידית ב-Positioned(left:0) יושבת על
+        // הדופן האמיתי (x=0 / x=רוחב), ולא 12px פנימה.
+        final paneRects = [
+          tester.getRect(panes.at(0)),
+          tester.getRect(panes.at(1)),
+        ];
+        final paneLeft =
+            paneRects.map((r) => r.left).reduce((a, b) => a < b ? a : b);
+        final paneRight =
+            paneRects.map((r) => r.right).reduce((a, b) => a > b ? a : b);
+
+        expect(paneLeft, 0,
+            reason: 'החלונית הקיצונית חייבת להיצמד לדופן החלון כדי שהידית '
+                '(Positioned(left:0)) תשב בדיוק על הדופן');
+        expect(paneRight, 1600,
+            reason: 'החלונית הקיצונית חייבת להיצמד לדופן החלון');
+
+        // תוכן הקריאה (AdaptiveSidePane) מוזרק 12px פנימה מדופן החלון.
+        final content = find.byType(AdaptiveSidePane);
+        expect(content, findsNWidgets(2));
+        final contentRects = [
+          tester.getRect(content.at(0)),
+          tester.getRect(content.at(1)),
+        ];
+        final contentLeft =
+            contentRects.map((r) => r.left).reduce((a, b) => a < b ? a : b);
+        final contentRight =
+            contentRects.map((r) => r.right).reduce((a, b) => a > b ? a : b);
+
+        expect(contentLeft, dividerWidth,
+            reason: 'תוכן הקריאה חייב לשמור 12px מדופן החלון; 0 פירושו '
+                'שה-SplitPaneContentInset לא הוזרק לתוכן');
+        expect(contentRight, 1600 - dividerWidth,
+            reason: 'תוכן הקריאה חייב לשמור 12px מדופן החלון');
+      },
+    );
+  });
+
   group('reading_screen — מניעת התנגשות נתיב PageStorage', () {
     Widget buildTree({
       required Key wrapperKey,
@@ -345,6 +461,61 @@ void main() {
       },
     );
   });
+}
+
+class _FakeSettingsBloc extends Bloc<SettingsEvent, SettingsState>
+    implements SettingsBloc {
+  _FakeSettingsBloc() : super(SettingsState.initial()) {
+    on<SettingsEvent>((_, __) {});
+  }
+
+  @override
+  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
+}
+
+class _FakePersonalNotesBloc
+    extends Bloc<PersonalNotesEvent, PersonalNotesState>
+    implements PersonalNotesBloc {
+  _FakePersonalNotesBloc()
+      : super(const PersonalNotesState(
+          isLoading: false,
+          bookId: '',
+          locatedNotes: [],
+          missingNotes: [],
+          errorMessage: null,
+          filteredLocatedNotes: [],
+          filteredMissingNotes: [],
+        )) {
+    on<PersonalNotesEvent>((_, __) {});
+  }
+
+  @override
+  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
+}
+
+class _FakeTabsBloc extends Bloc<TabsEvent, TabsState> implements TabsBloc {
+  _FakeTabsBloc(super.initial) {
+    on<TabsEvent>((_, __) {});
+  }
+
+  @override
+  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
+}
+
+/// כרטסיית מפרשים מזויפת — תוכן קל שנטען סינכרונית בתצוגה המפוצלת.
+PdfCommentatorsTab _commentaryTab(String title) {
+  final sourceTab = PdfBookTab(
+    book: PdfBook(title: title, path: '/tmp/$title.pdf'),
+    pageNumber: 1,
+  );
+  sourceTab.pdfHeadings = PdfHeadings(
+    bookTitle: title,
+    headingsMap: {'פרק א': 1},
+  );
+  sourceTab.currentTitle.value = 'פרק א';
+  sourceTab.currentTextLineNumber = 1;
+  sourceTab.currentTextLineNumberEnd = 9;
+  return PdfCommentatorsTab(sourceTab: sourceTab);
 }
 
 /// וידג'ט בדיקה שסופר כמה פעמים נקרא ה-initState שלו לכל label.
