+++ test/text_book/view/page_shape/page_shape_screen_settings_test.dart
@@ -1,10 +1,9 @@
-// טסט רגרסיה לפתיחת דיאלוג הגדרות צורת הדף דרך ה-notifier:
+// טסט רגרסיה לפתיחת פאנל הגדרות צורת הדף דרך ה-notifier:
 // כפתור גלגל השיניים בסרגל העליון (TextBookScreen) רק מסמן בקשה ב-notifier,
-// ו-PageShapeScreen הוא שפותח את הדיאלוג — עם onSettingsChanged מחווט, כדי
-// שכל שינוי בדיאלוג יוחל על המסך בעדכון חי בלי להמתין לסגירתו.
-// (רגרסיה: בעבר הכפתור פתח את הדיאלוג בעצמו בלי החיווט החי, והשינויים
-// הופיעו רק אחרי סגירת הדיאלוג.)
+// ו-PageShapeScreen הוא שפותח את פאנל ההגדרות הצף (ContextOverlayPanel) בצד
+// הימני — עם onSettingsChanged מחווט, כדי שכל שינוי יוחל על המסך בעדכון חי.
 
+import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_settings_screens/flutter_settings_screens.dart';
@@ -20,8 +19,9 @@ import 'package:otzaria/text_book/bloc/text_book_bloc.dart';
 import 'package:otzaria/text_book/bloc/text_book_event.dart';
 import 'package:otzaria/text_book/bloc/text_book_state.dart';
 import 'package:otzaria/text_book/view/page_shape/page_shape_screen.dart';
-import 'package:otzaria/text_book/view/page_shape/page_shape_settings_dialog.dart';
+import 'package:otzaria/text_book/view/page_shape/page_shape_settings_panel.dart';
 import 'package:otzaria/text_book/view/page_shape/utils/page_shape_settings_manager.dart';
+import 'package:otzaria/widgets/layout/context_overlay_panel.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
 
 import '../../../test_helpers/memory_cache_provider.dart';
@@ -33,8 +33,52 @@ void main() {
     await Settings.init(cacheProvider: MemoryCacheProvider());
   });
 
+  // פאנל ההגדרות הוא ה-ContextOverlayPanel היחיד במסך; בודקים את מצב
+  // הפתיחה דרך isOpen כי המעטפת נשארת בעץ גם כשהפאנל סגור.
+  ContextOverlayPanel settingsPane(WidgetTester tester) =>
+      tester.widget<ContextOverlayPanel>(find.byType(ContextOverlayPanel));
+
+  Future<void> pumpScreen(
+    WidgetTester tester,
+    ValueNotifier<int> openSettingsNotifier,
+  ) async {
+    await tester.pumpWidget(
+      MaterialApp(
+        home: MultiBlocProvider(
+          providers: [
+            BlocProvider<TextBookBloc>.value(
+              value: _TestTextBookBloc(_loadedState()),
+            ),
+            BlocProvider<PersonalNotesBloc>.value(
+              value: _TestPersonalNotesBloc(
+                const PersonalNotesState(
+                  isLoading: false,
+                  bookId: 'ספר בדיקה',
+                  locatedNotes: [],
+                  missingNotes: [],
+                  errorMessage: null,
+                  filteredLocatedNotes: [],
+                  filteredMissingNotes: [],
+                ),
+              ),
+            ),
+            BlocProvider<SettingsBloc>.value(
+              value: _TestSettingsBloc(SettingsState.initial()),
+            ),
+          ],
+          child: PageShapeScreen(
+            openBookCallback: (_) {},
+            openSettingsNotifier: openSettingsNotifier,
+          ),
+        ),
+      ),
+    );
+    await tester.pump();
+    await tester.pump();
+  }
+
   testWidgets(
-      'בקשת פתיחה דרך openSettingsNotifier פותחת את הדיאלוג עם עדכון חי מחווט',
+      'בקשת פתיחה דרך openSettingsNotifier פותחת את החלונית עם עדכון חי מחווט',
       (tester) async {
     final openSettingsNotifier = ValueNotifier<int>(0);
     addTearDown(openSettingsNotifier.dispose);
@@ -75,21 +119,91 @@ void main() {
     await tester.pump();
     await tester.pump();
 
-    expect(find.byType(PageShapeSettingsDialog), findsNothing);
+    expect(find.byType(PageShapeSettingsPanel), findsNothing);
 
     // לחיצה על כפתור גלגל השיניים בסרגל העליון מתורגמת להעלאת ה-notifier
     openSettingsNotifier.value++;
     await tester.pump();
     await tester.pump();
 
-    expect(find.byType(PageShapeSettingsDialog), findsOneWidget);
+    expect(find.byType(PageShapeSettingsPanel), findsOneWidget);
 
-    // החיווט החי: הדיאלוג חייב לקבל onSettingsChanged כדי שהמסך יתעדכן
-    // תוך כדי שינוי, ולא רק אחרי סגירת הדיאלוג.
-    final dialog = tester.widget<PageShapeSettingsDialog>(
-      find.byType(PageShapeSettingsDialog),
+    // החיווט החי: החלונית חייבת לקבל onSettingsChanged כדי שהמסך יתעדכן
+    // תוך כדי שינוי, ולא רק אחרי סגירתה.
+    final panel = tester.widget<PageShapeSettingsPanel>(
+      find.byType(PageShapeSettingsPanel),
     );
-    expect(dialog.onSettingsChanged, isNotNull);
+    expect(panel.onSettingsChanged, isNotNull);
+    // כפתור האיפוס חייב להיות מחווט כדי שהמסך יטען מחדש את ברירות המחדל.
+    expect(panel.onReset, isNotNull);
+    // כפתור ה-X של המעטפת חייב להיות מחווט לסגירת הפאנל.
+    expect(settingsPane(tester).onClose, isNotNull);
+  });
+
+  testWidgets('יריית notifier שנייה סוגרת את חלונית ההגדרות (טוגל)',
+      (tester) async {
+    final openSettingsNotifier = ValueNotifier<int>(0);
+    addTearDown(openSettingsNotifier.dispose);
+
+    await pumpScreen(tester, openSettingsNotifier);
+    expect(settingsPane(tester).isOpen, isFalse);
+
+    openSettingsNotifier.value++;
+    await tester.pump();
+    await tester.pump();
+    expect(settingsPane(tester).isOpen, isTrue);
+
+    // יריית notifier שנייה = טוגל → החלונית נסגרת.
+    openSettingsNotifier.value++;
+    await tester.pump();
+    await tester.pump();
+    expect(settingsPane(tester).isOpen, isFalse);
+  });
+
+  testWidgets('כפתור הסגירה בחלונית סוגר אותה וניתן לפתוח שוב', (tester) async {
+    final openSettingsNotifier = ValueNotifier<int>(0);
+    addTearDown(openSettingsNotifier.dispose);
+
+    await pumpScreen(tester, openSettingsNotifier);
+
+    openSettingsNotifier.value++;
+    await tester.pump();
+    // המתנה לסיום אנימציית ההחלקה — כפתור ה-X נחשף אחרון (בצד הפנימי)
+    // ולכן חייב שהחלונית תיפתח במלואה לפני הלחיצה.
+    await tester.pump(const Duration(milliseconds: 500));
+    expect(settingsPane(tester).isOpen, isTrue);
+
+    await tester.tap(
+      find.widgetWithIcon(IconButton, FluentIcons.dismiss_24_regular),
+    );
+    await tester.pump();
+    await tester.pump();
+    expect(settingsPane(tester).isOpen, isFalse);
+
+    // אחרי סגירה מכפתור ה-X עדיין ניתן לפתוח מחדש דרך ה-notifier.
+    openSettingsNotifier.value++;
+    await tester.pump();
+    await tester.pump();
+    expect(settingsPane(tester).isOpen, isTrue);
+  });
+
+  testWidgets('הקשה על ה-scrim מחוץ לפאנל סוגרת אותו', (tester) async {
+    final openSettingsNotifier = ValueNotifier<int>(0);
+    addTearDown(openSettingsNotifier.dispose);
+
+    await pumpScreen(tester, openSettingsNotifier);
+
+    openSettingsNotifier.value++;
+    await tester.pump();
+    await tester.pump(const Duration(milliseconds: 500));
+    expect(settingsPane(tester).isOpen, isTrue);
+    expect(find.byType(PageShapeSettingsPanel), findsOneWidget);
+
+    // הפאנל (רוחב 400) יושב בצד הימני; הקשה בקצה השמאלי פוגעת ב-scrim.
+    await tester.tapAt(const Offset(30, 300));
+    await tester.pump();
+    await tester.pump();
+    expect(settingsPane(tester).isOpen, isFalse);
   });
 
   testWidgets('כשל טעינת מפרש תחתון אינו שומר הסתרה גלובלית', (tester) async {
